Properly add subscriptions to authorization type
